Merge base

+++ app/src/main/java/com/kabouzeid/gramophone/ui/activities/tageditor/AbsTagEditorActivity.java
@@ -10,10 +10,10 @@ import android.graphics.BitmapFactory;
 import android.media.MediaScannerConnection;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
@@ -75,7 +75,7 @@ public abstract class AbsTagEditorActivity extends AbsBaseActivity {
     ImageView image;
     @BindView(R.id.header)
     LinearLayout header;
-    private int id;
+    private long id;
     private int headerVariableSpace;
     private int paletteColorPrimary;
     private boolean isInNoImageMode;
@@ -186,7 +186,7 @@ public abstract class AbsTagEditorActivity extends AbsBaseActivity {
     private void getIntentExtras() {
         Bundle intentExtras = getIntent().getExtras();
         if (intentExtras != null) {
-            id = intentExtras.getInt(EXTRA_ID);
+            id = intentExtras.getLong(EXTRA_ID);
         }
     }
 
@@ -394,16 +394,16 @@ public abstract class AbsTagEditorActivity extends AbsBaseActivity {
     }
 
     public static class ArtworkInfo {
-        public final int albumId;
+        public final long albumId;
         public final Bitmap artwork;
 
-        public ArtworkInfo(int albumId, Bitmap artwork) {
+        public ArtworkInfo(long albumId, Bitmap artwork) {
             this.albumId = albumId;
             this.artwork = artwork;
         }
     }
 
-    protected int getId() {
+    protected long getId() {
         return id;
     }
